refactor: organize tag route tests by scope
