+++ backend/tests/routes/tags/__init__.py
@@ -1,0 +1,1 @@
+"""Tag route tests"""
